document_processor/backend/src/database.py
import psycopg2
import psycopg2.extras
from . import config # Relative import within src package

def get_db_connection():
    """Establishes a connection to the PostgreSQL database."""
    conn = None
    try:
        conn = psycopg2.connect(
            host=config.DB_HOST,
            port=config.DB_PORT,
            dbname=config.DB_NAME,
            user=config.DB_USER,
            password=config.DB_PASSWORD
        )
        return conn
    except psycopg2.OperationalError as e:
        print(f"Error connecting to database: {e}")
        if conn:
            conn.close()
        raise # Re-raise the exception to be handled by the caller 