document_processor/backend/requirements.txt
Flask
requests
pandas
Flask-CORS
psycopg2-binary
rapidfuzz 